Refactor character data skills element IDs
Changed the IDs of the character data skills elements in the 'character/details.html.twig' file to reflect the specific skill name. This enhances the readability of the code and, by creating unique IDs, it allows for more specific selection and manipulation within the JavaScript files. A commenting out of a section concerning navigation buttons was also done in this commit.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>