Uncomment XMLHttpRequest checks in CharacterEditController
The code for checks on whether a request is an AJAX (XmlHttpRequest) is uncommented in the CharacterEditController. Furthermore, a new XMLHttpRequest header is added to our AJAX requests in Ajax.js and Ajax.ts. These changes contribute to better handling of AJAX requests on the server side, improving response performance and reliability.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>