Add maxCharacterCount field to PNPGroup entity
A new field named maxCharacterCount has been added to the PNPGroup entity and its corresponding migration file has been created. This attribute allows for the specification of a maximum number of characters for a group in PNP. The database schema has been updated accordingly with the newly created Doctrine migration.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>