Add group invitation functionality and templates
This commit introduces a new feature, group invitation. New templates for create, list, show, redeem, and error pages for PNPGroupInvitation, along with the necessary controller code are added. Supporting changes to make this feature functional are included, such as addition of an InvitationTools class, updates to EntityRepositories and changes to route annotations.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>