Add CharacterXP and EuroDollar fields to character details section
Implemented additional fields for 'CharacterXP' and 'EuroDollar' in the character detail view. This allows the user to edit, save, and display these character attributes depending upon the editability of the character. These fields have been added in the UI as well as persisted in the character creation stage in the backend.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>