Refactor character routes to ruleset routes
The route names related to 'character' category in the character class, character stat, and character stat category controllers have been refactored to 'ruleset'. This also reflects in the respective Twig templates. The changes provide a more accurate description of route functionality and align it with the ruleset context. Finally, the security configuration is updated to grant ROLE_RULESET users access to these updated paths.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>