Update PNPGroupController and PNPGroupType's form
Modified the PNPGroupController and the PNPGroupType's form to improve functionality. The redirect path on form completion has been corrected and a feature to set a custom success flash message in PNPGroupController has been added. Additionally, the form now includes an integer field to handle the maximum character count in the PNPGroupType.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>