Add Symfony mailer & update dependencies
Added the Symfony/mailer package to the project for email services and updated other relevant dependencies. This includes additions to the composer.json and composer.lock files, and changes to various configuration and environment files. These updates provide the resources to set up and utilize email functionalities in the future.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>