Add character leaving group functionality
The update introduces the ability for a character to leave their role-playing game (RPG) group. There is now a leave function in the PNPGroupController that removes the character from the group entity and the option to leave the group has been added to the character list view in the respective template. Meanwhile, a minor syntax adjustment is made in the CharacterController.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>