Add AbstractWebTest for easier testing and enhance password reset
This commit introduces a new `AbstractWebTest` which provides several helper methods to streamline testing processes. It also involves updates in `UserMailController` and related templates to improve the user password reset functionality. The enhancements provide a more streamlined and reliable user-password reset process, and unnecessary codes are purged for better readability and performance.

Signed-off-by: Jan-Peter Wittig <jan-peter.wittig@live.com>